% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingNA.R
\name{detectna}
\alias{detectna}
\title{Detect Missing Values}
\usage{
detectna(data)
}
\arguments{
\item{data}{A data frame or tibble}
}
\value{
A named vector with counts of missing values for each column
}
\description{
This function detects missing vales in each column of a dataset.
It return a named vector with counts of missing values for each column that has missing data
}
\examples{
detectna(airquality)
}
