% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repmissingNA.R
\name{repwmedian}
\alias{repwmedian}
\title{Replace missing values with median}
\usage{
repwmedian(dataset)
}
\arguments{
\item{dataset}{A data frame or tibble}
}
\value{
A data frame with missing values replaced by column median.
}
\description{
This function replaces missing values in numeric columns with the median of each column.
}
\examples{
repwmedian(airquality)
}
